+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 All notable changes to the Zowe Client Python SDK will be documented in this file.
 
+## Recent Changes
+
+- Feature: Added method to Save profile properties to zowe.config.json file